# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ARGUS-PRISM API
  version: 3.0.0
  description: |
    Pre-crime Intelligence System for Mule Detection — the single source of truth.

    **Conventions (enforced by CI, no exceptions):**
    - Success responses are enveloped: `{ "data": ..., "meta": {...} }`.
    - List responses put pagination in `meta`: `{ "cursor": ..., "total": ... }`.
    - Errors are RFC-7807 `application/problem+json`.
    - No response consumed by the UI ever contains key material (Law 3).
  contact:
    name: ARGUS-PRISM Engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development
  - url: /
    description: Same-origin (behind reverse proxy)
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token
      operationId: refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: New token pair (old refresh token revoked).
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/TokenPair" }
        "401": { $ref: "#/components/responses/Problem" }
...
  /api/v1/accounts:
    get:
      tags: [Accounts]
      summary: Account lookup (cursor-paginated, server-sorted)
      operationId: listAccounts
      parameters:
        - { name: query, in: query, schema: { type: string }, description: "Match account ref or holder." }
        - { name: risk_tier, in: query, schema: { $ref: "#/components/schemas/Severity" } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        "200":
          description: A page of account summaries.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountListEnvelope" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/accounts/{id}:
    get:
      tags: [Accounts]
      summary: Account profile (PII masked by role)
      operationId: getAccount
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Account profile.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Account" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/accounts/{id}/score-history:
    get:
      tags: [Accounts]
      summary: WarmthScore trajectory + SHAP per point
      operationId: getScoreHistory
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Score history.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ScorePoint" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/accounts/{id}/transactions:
    get:
      tags: [Accounts]
      summary: Account transactions (cursor-paginated)
      operationId: getAccountTransactions
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        "200":
          description: A page of transactions.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransactionListEnvelope" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/accounts/{id}/devices:
    get:
      tags: [Accounts]
      summary: IMEI/SIM event history
      operationId: getAccountDevices
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Device events.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/DeviceEvent" }
  /api/v1/accounts/{id}/signals:
    get:
      tags: [Accounts]
      summary: Current S1–S6 signal breakdown
      operationId: getAccountSignals
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Signal breakdown.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/SignalBreakdown" }
  /api/v1/accounts/{id}/actions:
    post:
      tags: [Accounts]
      summary: "Freeze / restrict / KYC-review (MLRO; server re-checks role)"
      operationId: accountAction
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountAction" }
      responses:
        "200":
          description: Updated account.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Account" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }

  # ─────────────────────────────── Network Graph ────────────────────────
  /api/v1/graph/neighborhood/{account_id}:
    get:
      tags: [Graph]
      summary: Force-directed neighborhood subgraph
      operationId: getNeighborhood
      parameters:
        - { name: account_id, in: path, required: true, schema: { type: string } }
        - { name: hops, in: query, schema: { type: integer, minimum: 1, maximum: 4, default: 3 } }
      responses:
        "200":
          description: Nodes + edges within N hops.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Graph" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/graph/patterns/{type}:
    get:
      tags: [Graph]
      summary: Detected layering / round-trip / structuring subgraphs
      operationId: getPatterns
      parameters:
        - { name: type, in: path, required: true, schema: { type: string, enum: [round_trip, layering, structuring, velocity, any] } }
      responses:
        "200":
          description: Matching pattern subgraphs.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PatternMatch" }
  /api/v1/graph/freeze-cluster:
    post:
      tags: [Graph]
      summary: "Freeze a selected node set (MLRO); propagates taint"
      operationId: freezeCluster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_ids]
              properties:
                account_ids: { type: array, items: { type: string } }
                reason: { type: string }
      responses:
        "200":
          description: Freeze result + taint summary.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/FreezeResult" }
        "403": { $ref: "#/components/responses/Problem" }

  # ─────────────────────────────── Recruiter Map ────────────────────────
...
  /api/v1/recruiters/{id}/campaign:
    get:
      tags: [Recruiter]
      summary: Full fan-out subgraph for a recruiter
      operationId: getRecruiterCampaign
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Campaign subgraph.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/RecruiterCampaign" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/recruiters/{id}/freeze-campaign:
    post:
      tags: [Recruiter]
      summary: "One-click network freeze (MLRO); propagates taint"
      operationId: freezeCampaign
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Freeze result + taint summary.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/FreezeResult" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }

  # ─────────────────────────────── Stream / Metrics ─────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Problem:
      description: RFC-7807 problem+json error.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Meta:
      type: object
      properties:
        cursor: { type: [string, "null"], description: "Opaque cursor for the next page; null when exhausted." }
        total: { type: [integer, "null"] }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, default: "about:blank" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: "Machine-readable error code, e.g. mfa_required, rate_limited." }

    # ── Health ────────────────────────────────────────────────
...
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, const: bearer }
        expires_in: { type: integer, description: "Access token TTL in seconds." }
...
    Severity: { type: string, enum: [CLEAN, WARMING, HOT, CRITICAL, IMMINENT] }
...
    ShapSignal:
      type: object
      required: [code, label, contribution]
      properties:
        code: { type: string, example: S3 }
        label: { type: string, example: "Round-trip ratio 0.94" }
        contribution: { type: number, description: "Warmth points contributed." }
...
    AccountSummary:
      type: object
      required: [id, account_ref, holder, warmth_score, severity, status, tainted]
      properties:
        id: { type: string }
        account_ref: { type: string, description: "Masked reference for lists." }
        holder: { type: string, description: "Holder name, masked unless the role may view PII." }
        branch: { type: string }
        segment: { type: string }
        warmth_score: { type: number }
        severity: { $ref: "#/components/schemas/Severity" }
        status: { $ref: "#/components/schemas/AccountStatus" }
        tainted: { type: boolean }
    Account:
      allOf:
        - $ref: "#/components/schemas/AccountSummary"
        - type: object
          required: [kyc_status, response_tier, opened_at, last_active]
          properties:
            ifsc: { type: string }
            kyc_status: { type: string }
            status_reason: { type: [string, "null"] }
            response_tier: { type: integer }
            taint_score: { type: number }
            opened_at: { type: string, format: date-time }
            last_active: { type: string, format: date-time }
            pii_masked: { type: boolean, description: "True when the caller's role only sees masked PII." }
            top_signals:
              type: array
              items: { $ref: "#/components/schemas/ShapSignal" }
    AccountStatus:
      type: string
      enum: [ACTIVE, WATCH, KYC_HOLD, RESTRICTED, FROZEN]
    AccountListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/AccountSummary" }
        meta: { $ref: "#/components/schemas/Meta" }
    ScorePoint:
      type: object
      required: [ts, score, severity]
      properties:
        ts: { type: string, format: date-time }
        score: { type: number }
        severity: { $ref: "#/components/schemas/Severity" }
        shap:
          type: array
          items: { $ref: "#/components/schemas/ShapSignal" }
    AccountTransaction:
      type: object
      required: [id, ts, amount, direction, channel]
      properties:
        id: { type: string }
        ts: { type: string, format: date-time }
        counterparty_ref: { type: [string, "null"], description: "Masked counterparty, or null if external." }
        amount: { type: number }
        direction: { type: string, enum: [IN, OUT] }
        channel: { type: string }
        description: { type: [string, "null"] }
    TransactionListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/AccountTransaction" }
        meta: { $ref: "#/components/schemas/Meta" }
    DeviceEvent:
      type: object
      required: [imei, event_type, registered_at]
      properties:
        imei: { type: string, description: "Masked IMEI (last 4 shown)." }
        event_type: { type: string }
        registered_at: { type: string, format: date-time }
    SignalBreakdown:
      type: object
      required: [score, signals, shap]
      properties:
        score: { type: number }
        signals:
          type: object
          additionalProperties: { type: number }
          description: "Raw S1–S6 values in [0,1]."
        shap:
          type: array
          items: { $ref: "#/components/schemas/ShapSignal" }
    AccountAction:
      type: object
      required: [action]
      properties:
        action: { type: string, enum: [freeze, restrict, kyc_review, unfreeze] }
        reason: { type: string }

    # ── Graph ─────────────────────────────────────────────────
    GraphNode:
      type: object
      required: [id, account_ref, warmth_score, severity, tainted]
      properties:
        id: { type: string }
        account_ref: { type: string }
        holder: { type: string }
        warmth_score: { type: number }
        severity: { $ref: "#/components/schemas/Severity" }
        status: { $ref: "#/components/schemas/AccountStatus" }
        tainted: { type: boolean }
        is_recruiter: { type: boolean }
    GraphEdge:
      type: object
      required: [source, target, value, count]
      properties:
        source: { type: string }
        target: { type: string }
        value: { type: number, description: "Summed transaction value (edge weight)." }
        count: { type: integer }
    Graph:
      type: object
      required: [root, nodes, edges]
      properties:
        root: { type: string }
        nodes: { type: array, items: { $ref: "#/components/schemas/GraphNode" } }
        edges: { type: array, items: { $ref: "#/components/schemas/GraphEdge" } }
    PatternMatch:
      type: object
      required: [anchor, anchor_id, severity, nodes, edges]
      properties:
        anchor: { type: string }
        anchor_id: { type: string }
        severity: { $ref: "#/components/schemas/Severity" }
        warmth_score: { type: number }
        nodes: { type: array, items: { $ref: "#/components/schemas/GraphNode" } }
        edges: { type: array, items: { $ref: "#/components/schemas/GraphEdge" } }
    FreezeResult:
      type: object
      required: [frozen, tainted]
      properties:
        frozen: { type: integer, description: "Accounts frozen." }
        tainted: { type: integer, description: "Downstream accounts tainted by propagation." }
        account_refs: { type: array, items: { type: string } }

    # ── Recruiter ─────────────────────────────────────────────
...
    RecruiterCampaign:
      type: object
      required: [recruiter, mule_ids, nodes, edges]
      properties:
        recruiter:
          type: object
          properties:
            id: { type: string }
            account_ref: { type: string }
            scale_class: { type: string }
            downstream_count: { type: integer }
        mule_ids: { type: array, items: { type: string } }
        nodes: { type: array, items: { $ref: "#/components/schemas/GraphNode" } }
        edges: { type: array, items: { $ref: "#/components/schemas/GraphEdge" } }

    # ── Simulator ─────────────────────────────────────────────
...
security:
  - bearerAuth: []